RNG now attempts to recover when not random enough
Extends error handler to reset the peripheral. Specifically should solve issues where there is a seed error in the RNG. We just reset the whole peripheral which is definitely overkill for some errors, but I think realistically the only thing we've seen is a seed error where the values continue to return a loop without enough entropy.